pyproject.toml
[tool.poetry]
name = "test-qortex"
version = "0.0.1"
description = "API каталога исполнителей с альбомами и их песнями."
authors = [
    "Никита Шидогубов",
]

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.1.1"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
drf-spectacular = "^0.28.0"
django-filter = "^25.1"
faker = "^37.6.0"
coverage = "^7.10.6"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.3.0"
black = "^25.1.0"
ruff = "^0.12.11"


[tool.poetry.group.test.dependencies]

[tool.ruff]
line-length = 119
exclude = [
    ".git",
    ".gitignore",
    "__pycache__",
    "__init__.py",
    "manage.py",
    "poetry.lock",
    "pyproject.toml",
    "migrations",
    "tests.py",
]
target-version = "py312"
fix = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
ignore = ["E501", "W293"]
select = ["E", "F", "I", "B", "C", "A", "N", "UP"]
mccabe.max-complexity = 12

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

